created delete and edit files. made some javascript

+++ app/posts/delete.php
@@ -1,20 +1,25 @@
 <?php
-
 declare(strict_types=1);
-
 require __DIR__.'/../autoload.php';
-
-$deletePost = filter_var(trim($_GET['user']),FILTER_SANITIZE_STRING);
-$statement = $pdo->prepare('SELECT user_id FROM posts WHERE id = :id');
-$statement->bindParam(':id', $deletePost, PDO::PARAM_STR);
-$statement->execute();
-
-$postId= $statement->fetch(PDO::FETCH_ASSOC);
-
+// In this file we delete new posts in the database.
+if(isset($_POST['delete_post'])){
+$delete = $_POST['delete_post'];
 
 $statement = $pdo->prepare('DELETE FROM posts WHERE id = :id');
-$statement->bindParam(':id', $deletePost, PDO::PARAM_INT);
+//if not die
+if (!$statement)
+{
+    die(var_dump($pdo->errorInfo()));
+}
+$statement->bindParam(':id', $delete, PDO::PARAM_INT);
 $statement->execute();
 
+$_SESSION['posts'] = [
 
-redirect('/');
+    'user_id' => $id,
+    'content' => $content,
+    'description' => $description,
+    'created_at' => $created,
+];
+}
+redirect('/profile.php');
